# webapi/openapi.generated.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "dixcord-bot webapi",
    "version": "3.0.0",
    "description": "Internal REST API backing the dixcord-bot Discord bot, webview, and webadmin services. Every route requires a service-account JWT (admin, bot, or webview) except login and health."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Service-account JWT obtained from POST /api/auth/login."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "ok",
          "error"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "string"
          }
        }
      },
      "AuthErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "description": "Shape returned by the authenticate/requireAdmin middleware, which omits `ok`.",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing, invalid, or expired bearer token.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AuthErrorResponse"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Resource not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Request body or query parameters failed validation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ServerError": {
        "description": "Unexpected server error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/guild": {
      "get": {
        "operationId": "getGuild",
        "tags": [
          "Guild"
        ],
        "summary": "Get the synced guild snapshot",
        "description": "Returns the most recently synced (app, guildId) pair when neither query param is given.",
        "parameters": [
          {
            "name": "app",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "example": "discord"
            }
          },
          {
            "name": "guildId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Guild snapshot.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "guild": {
                      "type": "object"
                    },
                    "channels": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "roles": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "emojis": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "stickers": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "syncedAt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/guild/sync": {
      "post": {
        "operationId": "syncGuild",
        "tags": [
          "Guild"
        ],
        "summary": "Push a full guild snapshot",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "app",
                  "guildId",
                  "guild",
                  "channels",
                  "roles"
                ],
                "properties": {
                  "app": {
                    "type": "string",
                    "example": "discord"
                  },
                  "guildId": {
                    "type": "string"
                  },
                  "guild": {
                    "type": "object"
                  },
                  "channels": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  },
                  "roles": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Snapshot stored.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/guild-members/sync": {
      "post": {
        "operationId": "syncGuildMembers",
        "tags": [
          "Guild Members"
        ],
        "summary": "Push a full server membership list",
        "description": "Full-replace: deletes and reinserts every guild_members row for this (app, guildId). Entries whose platformUserId isn't yet known to chat_member_mapping are skipped (best-effort) rather than failing the whole sync.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "app",
                  "guildId",
                  "members"
                ],
                "properties": {
                  "app": {
                    "type": "string",
                    "example": "discord"
                  },
                  "guildId": {
                    "type": "string"
                  },
                  "members": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "platformUserId"
                      ],
                      "properties": {
                        "platformUserId": {
                          "type": "string"
                        },
                        "nickname": {
                          "type": "string",
                          "nullable": true
                        },
                        "roles": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "description": "guild_roles.id values held in this server."
                        },
                        "joinedAt": {
                          "type": "string",
                          "format": "date-time",
                          "nullable": true
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sync result.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "imported": {
                      "type": "integer"
                    },
                    "skipped": {
                      "type": "integer",
                      "description": "Entries skipped because platformUserId wasn't found in chat_member_mapping."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/api/guild-members": {
      "get": {
        "operationId": "listGuildMembers",
        "tags": [
          "Guild Members"
        ],
        "summary": "List one server's members",
        "parameters": [
          {
            "name": "app",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "example": "discord"
            }
          },
          {
            "name": "guildId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Members of this server.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "members": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "description": "chat_member_mapping.id"
                          },
                          "name": {
                            "type": "string"
                          },
                          "handle": {
                            "type": "string"
                          },
                          "platformUserId": {
                            "type": "string"
                          },
                          "nickname": {
                            "type": "string",
                            "nullable": true
                          },
                          "roles": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "joinedAt": {
                            "type": "string",
                            "nullable": true
                          },
                          "syncedAt": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  }
}